add pom pom and pom macros refactor code fixes #8 fixes #2
