+++ package.json
@@ -20,9 +20,12 @@
   },
   "homepage": "https://github.com/andreidimaano/Calcifer#readme",
   "devDependencies": {
-    "@types/node": "^14.14.10",
     "@types/dotenv": "^8.2.0",
+    "@types/jest": "^26.0.19",
+    "@types/node": "^14.14.10",
+    "jest": "^26.6.3",
     "nodemon": "^2.0.6",
+    "ts-jest": "^26.4.4",
     "ts-node": "^9.0.0",
     "ts-node-dev": "^1.0.0",
     "typescript": "^4.1.2"
