+++ src/command/pomodoro/pomodoro.ts
@@ -1,39 +1,6 @@
-import { Message, User, MessageEmbed } from "discord.js";
-
-let startEmbed = (time: number) => {
-    return new MessageEmbed()
-    .setColor('#dc2f02')
-    .setTitle('Pomodoro')
-    .setTimestamp()
-    .addFields(
-        { name: `:tomato: Your timer is set to ${time} minutes :tomato:`, value: ':blush: Happy Studying! :blush:'},
-    )
-}
-
-let endEmbed = new MessageEmbed()
-    .setColor('#dc2f02')
-    .setTitle('Pomodoro')
-    .setTimestamp()
-    .addFields(
-        { name: `:tomato: Congrats on finishing a pomodoro session :tomato:`, value: ':blush: enjoy your break! :blush:' }
-    )
-
-let removeMember = (array: string[], author_tag: string) :string[] => {
-    //console.log('function called');
-    let index = array.indexOf(author_tag);
-    if(index > -1) {
-        return array.splice(index, 1);
-    }
-
-    return array;
-};
-
-let currentlyStudying = (author_tag: string) :boolean => {
-    return (currentMembersStudying.find(member_tag => member_tag === author_tag)) ? true : false;
-};
-
-let currentMembersStudying: string[] = [];
-
+import { Message } from "discord.js";
+import { endEmbed, startEmbed } from "./PomodoroEmbed";
+import { currentlyStudying, currentMembersStudying, removeMember } from "./PomodoroMembers";
 
 export let Pomodoro = async ( message: Message, time?: number) => {
     let timer = (time && time <= 50 && time >= 10) ? time : 25
@@ -53,7 +20,7 @@ export let Pomodoro = async ( message: Message, time?: number) => {
     setTimeout(async () => {
         await message.channel.send(message.author, endEmbed);
         removeMember(currentMembersStudying, message.author.tag);
-    }, /*60000*/ 10 * timer);
+    }, /*60000*/ 1000 * timer);
 }
